updated the poc wrt PR already merged in the release regarding the auto height container
