+++ app/client/src/components/autoHeight/AutoHeightContainer.tsx
@@ -1,10 +1,12 @@
-import React, { PropsWithChildren, useRef, useEffect, useState } from "react";
+import React, { PropsWithChildren, useEffect, useRef, useState } from "react";
 import { GridDefaults } from "constants/WidgetConstants";
 import styled from "styled-components";
+import { WidgetProps } from "widgets/BaseWidget";
 
 const StyledAutoHeightContainer = styled.div<{ isOverflow?: boolean }>`
   overflow-y: ${(props) => (props.isOverflow ? "auto" : "unset")};
   overflow-x: ${(props) => (props.isOverflow ? "hidden" : "unset")};
+  padding-right: 4px;
 `;
 
 interface AutoHeightContainerProps {
@@ -12,10 +14,11 @@ interface AutoHeightContainerProps {
   minDynamicHeight: number;
   isAutoHeightWithLimits: boolean;
   onHeightUpdate: (height: number) => void;
+  widgetProps?: WidgetProps;
 }
 
 const SimpleContainer = styled.div`
-  height: auto;
+  height: auto !important;
 `;
 
 export default function AutoHeightContainer({
@@ -24,12 +27,13 @@ export default function AutoHeightContainer({
   maxDynamicHeight,
   minDynamicHeight,
   onHeightUpdate,
+  widgetProps,
 }: PropsWithChildren<AutoHeightContainerProps>) {
   const [expectedHeight, setExpectedHeight] = useState(0);
 
   const ref = useRef<HTMLDivElement>(null);
 
-  const observer = useRef(
+  const observer = React.useRef(
     new ResizeObserver((entries) => {
       const height = entries[0].contentRect.height;
       setExpectedHeight(height);
@@ -58,10 +62,13 @@ export default function AutoHeightContainer({
       expectedHeight / GridDefaults.DEFAULT_GRID_ROW_HEIGHT,
     );
 
+    const backgroundColor = widgetProps?.backgroundColor;
+
     return (
       <StyledAutoHeightContainer
         className="auto-height-scroll-container"
         isOverflow={maxDynamicHeight < expectedHeightInRows}
+        style={{ backgroundColor }}
       >
         <SimpleContainer className="auto-height-container" ref={ref}>
           {children}
