+++ app/client/src/components/autoHeight/AutoHeightContainerWrapper.tsx
@@ -1,21 +1,20 @@
-import { WidgetProps } from "./BaseWidget";
+import React, { ReactNode } from "react";
+import useWidgetConfig from "utils/hooks/useWidgetConfig";
+import { DynamicHeight } from "utils/WidgetFeatures";
+import { WidgetProps } from "widgets/BaseWidget";
 import {
   getWidgetMaxDynamicHeight,
   getWidgetMinDynamicHeight,
-} from "./WidgetUtils";
-import DynamicHeightContainer from "./DynamicHeightContainer";
-import React, { ReactNode } from "react";
-import useWidgetConfig from "utils/hooks/useWidgetConfig";
+} from "widgets/WidgetUtils";
+import AutoHeightContainer from "./AutoHeightContainer";
 
-export type DynamicHeightWrapperProps = {
+export type AutoHeightWrapperProps = {
   widgetProps: WidgetProps;
   children: ReactNode;
   onUpdateDynamicHeight: (height: number) => void;
 };
 
-export function DynamicHeightContainerWrapper(
-  props: DynamicHeightWrapperProps,
-) {
+function AutoHeightContainerWrapper(props: AutoHeightWrapperProps) {
   const { children, widgetProps } = props;
   const isCanvas = useWidgetConfig(widgetProps.type, "isCanvas");
   // eslint-disable-next-line react/jsx-no-useless-fragment
@@ -30,15 +29,20 @@ export function DynamicHeightContainerWrapper(
   const maxDynamicHeight = getWidgetMaxDynamicHeight(widgetProps);
   const minDynamicHeight = getWidgetMinDynamicHeight(widgetProps);
 
+  const isAutoHeightWithLimits =
+    widgetProps.dynamicHeight === DynamicHeight.AUTO_HEIGHT_WITH_LIMITS;
+
   return (
-    <DynamicHeightContainer
-      dynamicHeight={widgetProps.dynamicHeight}
+    <AutoHeightContainer
+      isAutoHeightWithLimits={isAutoHeightWithLimits}
       maxDynamicHeight={maxDynamicHeight}
       minDynamicHeight={minDynamicHeight}
       onHeightUpdate={onHeightUpdate}
       widgetProps={widgetProps}
     >
       {children}
-    </DynamicHeightContainer>
+    </AutoHeightContainer>
   );
 }
+
+export default AutoHeightContainerWrapper;
